Attempt at Week 4 Assignment EXTREMELY MESSY CODE :(
game object can only fire a projectile once every 3 seconds
if the player hits the sides or collides on the sides of the game object, the player dies/ game resets (reload the scene).
the game object can move left to right.
he game object detects that there is no longer a platform ahead it will turn back and move in the other direction (NOT COMPLETE)